Refactor Label- and Unit- DeclFindUsagesHandler
